move DeviceTokensForUser into device_token.go
